+++ flink-ml-core/src/main/java/org/apache/flink/ml/common/datastream/DataStreamUtils.java
@@ -104,8 +104,7 @@ public class DataStreamUtils {
     }
 
     /**
-     * Applies a {@link MapPartitionFunction} on a bounded data stream. The type of the output is
-     * inferred automatically.
+     * Applies a {@link MapPartitionFunction} on a bounded data stream.
      *
      * @param input The input data stream.
      * @param func The user defined mapPartition function.
@@ -115,7 +114,9 @@ public class DataStreamUtils {
      */
     public static <IN, OUT> DataStream<OUT> mapPartition(
             DataStream<IN> input, MapPartitionFunction<IN, OUT> func) {
-        return mapPartition(input, func, null);
+        TypeInformation<OUT> outType =
+                TypeExtractor.getMapPartitionReturnTypes(func, input.getType(), null, true);
+        return mapPartition(input, func, outType);
     }
 
     /**
@@ -132,17 +133,13 @@ public class DataStreamUtils {
             DataStream<IN> input,
             MapPartitionFunction<IN, OUT> func,
             TypeInformation<OUT> outType) {
-        if (outType == null) {
-            outType = TypeExtractor.getMapPartitionReturnTypes(func, input.getType(), null, true);
-        }
         return input.transform("mapPartition", outType, new MapPartitionOperator<>(func))
                 .setParallelism(input.getParallelism());
     }
 
     /**
      * Applies a {@link ReduceFunction} on a bounded data stream. The output stream contains at most
-     * one stream record and its parallelism is one. The type of the output is inferred
-     * automatically.
+     * one stream record and its parallelism is one.
      *
      * @param input The input data stream.
      * @param func The user defined reduce function.
@@ -150,7 +147,7 @@ public class DataStreamUtils {
      * @return The result data stream.
      */
     public static <T> DataStream<T> reduce(DataStream<T> input, ReduceFunction<T> func) {
-        return reduce(input, func, null);
+        return reduce(input, func, input.getType());
     }
 
     /**
@@ -165,9 +162,6 @@ public class DataStreamUtils {
      */
     public static <T> DataStream<T> reduce(
             DataStream<T> input, ReduceFunction<T> func, TypeInformation<T> outType) {
-        if (outType == null) {
-            outType = input.getType();
-        }
         DataStream<T> partialReducedStream =
                 input.transform("reduce", outType, new ReduceOperator<>(func))
                         .setParallelism(input.getParallelism());
@@ -182,7 +176,7 @@ public class DataStreamUtils {
 
     /**
      * Applies a {@link ReduceFunction} on a bounded keyed data stream. The output stream contains
-     * one stream record for each key. The type of the output is inferred automatically.
+     * one stream record for each key.
      *
      * @param input The input keyed data stream.
      * @param func The user defined reduce function.
@@ -191,7 +185,7 @@ public class DataStreamUtils {
      * @return The result data stream.
      */
     public static <T, K> DataStream<T> reduce(KeyedStream<T, K> input, ReduceFunction<T> func) {
-        return reduce(input, func, null);
+        return reduce(input, func, input.getType());
     }
 
     /**
@@ -207,9 +201,6 @@ public class DataStreamUtils {
      */
     public static <T, K> DataStream<T> reduce(
             KeyedStream<T, K> input, ReduceFunction<T> func, TypeInformation<T> outType) {
-        if (outType == null) {
-            outType = input.getType();
-        }
         return input.transform(
                         "Keyed Reduce",
                         outType,
@@ -221,8 +212,7 @@ public class DataStreamUtils {
     /**
      * Aggregates the elements in each partition of the input bounded stream, and then merges the
      * partial results of all partitions. The output stream contains the aggregated result and its
-     * parallelism is one. The types of the accumulated values and the output are inferred
-     * automatically.
+     * parallelism is one.
      *
      * <p>Note: If the parallelism of the input stream is N, this method would invoke {@link
      * AggregateFunction#createAccumulator()} N times and {@link AggregateFunction#merge(Object,
@@ -239,7 +229,13 @@ public class DataStreamUtils {
      */
     public static <IN, ACC, OUT> DataStream<OUT> aggregate(
             DataStream<IN> input, AggregateFunction<IN, ACC, OUT> func) {
-        return aggregate(input, func, null, null);
+        TypeInformation<ACC> accType =
+                TypeExtractor.getAggregateFunctionAccumulatorType(
+                        func, input.getType(), null, true);
+        TypeInformation<OUT> outType =
+                TypeExtractor.getAggregateFunctionReturnType(func, input.getType(), null, true);
+
+        return aggregate(input, func, accType, outType);
     }
 
     /**
@@ -267,15 +263,6 @@ public class DataStreamUtils {
             AggregateFunction<IN, ACC, OUT> func,
             TypeInformation<ACC> accType,
             TypeInformation<OUT> outType) {
-        if (accType == null) {
-            accType =
-                    TypeExtractor.getAggregateFunctionAccumulatorType(
-                            func, input.getType(), null, true);
-        }
-        if (outType == null) {
-            outType =
-                    TypeExtractor.getAggregateFunctionReturnType(func, input.getType(), null, true);
-        }
         DataStream<ACC> partialAggregatedStream =
                 input.transform(
                         "partialAggregate", accType, new PartialAggregateOperator<>(func, accType));
