+++ flink-ml-iteration/src/test/java/org/apache/flink/iteration/compile/DraftExecutionEnvironmentSwitchWrapperTest.java
@@ -54,7 +54,8 @@ public class DraftExecutionEnvironmentSwitchWrapperTest extends TestLogger {
     @Test
     public void testSwitchingOperatorWrappers() {
         StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
-        env.getConfig().enableObjectReuse().disableGenericTypes();
+        env.getConfig().enableObjectReuse();
+        env.getConfig().disableGenericTypes();
         DataStream<Integer> source =
                 env.addSource(new DraftExecutionEnvironment.EmptySource<Integer>() {});
