+++ flink-ml-lib/src/test/java/org/apache/flink/ml/util/TestUtils.java
@@ -52,7 +52,8 @@ public class TestUtils {
         Configuration config = new Configuration();
         config.set(ExecutionCheckpointingOptions.ENABLE_CHECKPOINTS_AFTER_TASKS_FINISH, true);
         StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment(config);
-        env.getConfig().enableObjectReuse().disableGenericTypes();
+        env.getConfig().enableObjectReuse();
+        env.getConfig().disableGenericTypes();
         env.setParallelism(4);
         env.enableCheckpointing(100);
         env.setRestartStrategy(RestartStrategies.noRestart());
